RAG-demo/requirements.txt
faiss-cpu>=1.13.0
langgraph>=1.0.0
langsmith>=0.7.0
litellm>=1.80.0
pypdf>=6.0.0
python-dotenv>=1.0.0
sentence-transformers>=5.0.0
torch>=2.10.0
